Move serialization to handler
