+++ vespa-osgi-testrunner/src/main/java/com/yahoo/vespa/testrunner/TestRunner.java
@@ -17,5 +17,4 @@ public interface TestRunner {
 
     TestReport getReport();
 
-    String getReportAsJson();
 }
