+++ vespa-osgi-testrunner/src/test/java/com/yahoo/vespa/testrunner/TestRunnerHandlerTest.java
@@ -145,9 +145,5 @@ public class TestRunnerHandlerTest {
             return testReport;
         }
 
-        @Override
-        public String getReportAsJson() {
-            return getReport().toJson();
-        }
     }
 }
